# lib/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Jambonz REST API
  description: Jambonz REST API specification
  contact:
    email: daveh@drachtio.org
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
  - url: /v1
    description: jambonz API server
...
paths:
...
  /InviteCodes:
    post:
      summary: validate an invite code
      operationId: validateInviteCode
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
              required:
                - code
      responses:
        204:
          description: code successfully validated
        404:
          description: code not found
...
  /signin:
    post:
      tags:
        - Authentication
      summary: sign in using email and password
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: successfully signed in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
              example:
                user_sid: 3192af15-8260-439f-939d-184613f6bbfb
                account_sid: 7be2d35a-da6e-4faf-a382-1b37a571c7fd
                name: Dave Horton
                email: daveh@drachtio.org
                provider: github
                provider_userid: davehorton
                pristine: true
                account_validated: true
                scope: read-write
        403:
          description: Invalid password
        404:
          description: email not found
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
  /login:
    post:
      tags:
        - Authentication
      summary: login and retrieve a JWT
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
      responses:
        200:
          description: user logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessfulLogin' 
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
  /logout:
    post:
      tags:
        - Authentication
      summary: log out and deactivate the JWT
      operationId: logoutUser
      responses:
        204:
          description: user logged out
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
  /forgot-password:
    post:
      tags:
        - Authentication
      summary: send link to reset password
      operationId: forgotPassword
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
      responses:
        204:
          description: email sent
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
              example:
                msg: service_provider_sid is missing
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
...
  /register:
    post:
      tags:
        - Authentication
      summary: create a new user and account
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                service_provider_sid:
                  type: string
                  example: 2708b1b3-2736-40ea-b502-c53d8396247f
                provider:
                  type: string
                  enum:
                    - github
                    - google
                    - twitter
                    - local
                  example: github
                email:
                  type: string
                password:
                  type: string
                email_activation_code:
                  type: string
                oauth2_code:
                  type: string
                  example: f82659563e061e7347de
                oauth2_state:
                  type: string
                  example: 386d2e990ad
                oauth2_client_id:
                  type: string
                  example: a075a5889264b8fbc831
                oauth2_redirect_uri:
                  type: string
                  example: https://localhost:3000/oauth-gh-callback
              required:
                - service_provider_sid
                - provider
      responses:
        200:
          description: user and account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
              example:
                user_sid: 3192af15-8260-439f-939d-184613f6bbfb
                account_sid: 7be2d35a-da6e-4faf-a382-1b37a571c7fd
                name: Dave Horton
                email: daveh@drachtio.org
                provider: github
                provider_userid: davehorton
                pristine: true
                account_validated: true
                scope: read-write

        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
              example:
                msg: service_provider_sid is missing
        422:
          description: User exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
              example:
                msg: invalid service_provider_sid   
        403:
          description: invalid/expired oauth2 code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
              example:
                msg: code is expired or invalid
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
...
  /Accounts/{AccountSid}/Conferences:
    get:
      tags:
       - Conferences
      summary: list conferences
      operationId: listConferences
      parameters:
      - name: AccountSid
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: list of conferences for a specified account
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
  /Accounts/{AccountSid}/Calls:
    post:
      tags:
      - Accounts
      summary: create a call
      operationId: createCall
      parameters:
      - name: AccountSid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              required:
                - from
                - to
              type: object
              properties:
                application_sid:
                  type: string
                  format: uuid
                  description: The application to use to control this call.  Either applicationSid or url is required.
                answerOnBridge:
                  type: boolean
                  description: If set to true, the inbound call will ring until the number that was dialed answers the call, and at that point a 200 OK will be sent on the inbound leg.  If false, the inbound call will be answered immediately as the outbound call is placed.
                  example: false
                call_hook:
                  $ref: '#/components/schemas/Webhook'
                  example: {"url": "https://acme.com/callhook", "method": "POST"}
                call_status_hook:
                  $ref: '#/components/schemas/Webhook'
                  example: {"url": "https://acme.com/status", "method": "POST"}
                from:
                  type: string
                  description: The calling party number
                  example: "16172375089"
                fromHost:
                  type: string
                  description: The hostname to put in the SIP From header of the INVITE
                  example: "blf.finotel.com"
                timeout:
                  type: integer
                  description: The number of seconds to wait for call to be answered.  Defaults to 60.
                  example: 30
                timeLimit:
                  type: integer
                  description: The max length of call in seconds
                  example: 60
                tag:
                  type: object
                  description: Initial set of customer-supplied metadata to associate with the call (see jambonz 'tag' verb)
                  example: {"callCount": 10}
                to:
                  $ref: '#/components/schemas/Target'
                  description: Destination for call
                headers:
                  type: object
                  description: The customer SIP headers to associate with the call
                  example: {"X-Custom-Header": "Hello"}
                sipRequestWithinDialogHook:
                  type: string
                  description: The sip indialog hook to receive session messages
                  example: '/customHook'
      responses:
        201:
          description: call successfully created
          content:
            application/json:
              schema:
                required:
                  - sid
                properties:
                  sid:
                    type: string
                    format: uuid
                    example: 2531329f-fb09-4ef7-887e-84e648214436
        400:
          description: bad request
    get:
      tags:
       - Accounts
      summary: list calls
      operationId: listCalls
      parameters:
      - name: AccountSid
        in: path
        required: true
        schema:
          type: string
      - in: query
        name: direction
        required: false
        schema:
          type: string
          enum:
            - inbound
            - outbound
        description: call direction to retrieve
      - in: query
        name: from
        required: false
        schema:
          type: string
        description: calling number to retrieve
      - in: query
        name: to
        required: false
        schema:
          type: string
        description: called number to retrieve
      - in: query
        name: callStatus
        required: false
        schema:
          type: string
          enum:
            - trying
            - ringing
            - early-media
            - in-progress
            - completed
            - failed
            - busy
            - no-answer
            - queued
        description: call status to retrieve
      responses:
        200:
          description: list of calls for a specified account
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Call'
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
  /Accounts/{AccountSid}/Calls/{CallSid}:
    parameters:
      - name: AccountSid
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: CallSid
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
    delete:
      tags:
       - Accounts
      summary: delete a call
      operationId: deleteCall
      responses:
        204:
          description: call successfully deleted
        404:
          description: call not found
        422:
          description: unprocessable entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
    get:
      tags:
       - Accounts
      summary: retrieve a call
      operationId: getCall
      responses:
        200:
          description: call found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Call'
        404:
          description: call not found
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
    post:
      tags:
       - Accounts
      summary: update a call
      operationId: updateCall
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                call_hook:
                  $ref: '#/components/schemas/Webhook'
                child_call_hook:
                  $ref: '#/components/schemas/Webhook'
                call_status:
                  type: string
                  enum:
                    - completed
                    - no-answer
                conf_mute_status:
                  type: string
                  enum:
                    - mute
                    - unmute
                conf_hold_status:
                  type: string
                  enum:
                    - hold
                    - unhold
                listen_status:
                  type: string
                  enum:
                    - pause
                    - silence
                    - resume
                listen:
                  type: object
                  description: >
                    Start or stop an ad-hoc audio fork (listen) on the call, without
                    requiring a listen/stream verb to have been declared up front.
                  properties:
                    action:
                      type: string
                      enum:
                        - start
                        - stop
                    url:
                      type: string
                      description: websocket url to stream audio to; required when action is 'start'
                mute_status:
                  type: string
                  enum:
                    - mute
                    - unmute
                whisper:
                  $ref: '#/components/schemas/Webhook'
                sip_request:
                  type: object
                  properties:
                    method:
                      type: string
                    content_type:
                      type: string
                    content:
                      type: string
                    headers:
                      type: object
                record: 
                  type: object
                  properties:
                    action:
                      type: string
                      enum:
                      - startCallRecording
                      - stopCallRecording
                      - pauseCallRecording
                      - resumeCallRecording
                    recordingID:
                      type: string 
                    siprecServerURL:
                      type: string
                conferenceParticipantAction:
                  type: object
                  properties:
                    action:
                      type: string
                      enum:
                        - tag
                        - untag
                        - coach
                        - uncoach
                        - mute
                        - unmute
                        - hold
                        - unhold
                    tag:
                      type: string
      responses:
        200:
          description: Accepted
        202:
          description: Accepted
        400:
          description: bad request
        404:
          description: call not found
        422:
          description: unprocessable entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  schemas:
    SuccessfulLogin:
      type: object
      required:
        - username
        - password
      properties:
        token: 
          type: string
        user_sid:
          type: string
        scope:
          type: string
        force_change:
          type: boolean
    Login:
      type: object
      properties:
        username: 
          type: string
        password:
          type: string        
      required:
        - username
        - password
...
    GeneralError:
      type: object
      required:
        - msg
      properties:
        msg:
          type: string
      example:
        msg: specific error detail will be provided here
...
    Webhook:
      type: object
      properties:
        webhook_sid:
          type: string
          format: uuid
        url:
          type: string
          format: url
        method:
          type: string
          enum:
            - get
            - post
        username:
          type: string
        password:
          type: string
      required:
        - url
      example: {"url": "https://acme.com", "method": "POST"}
...
    Call:
      type: object
      properties:
        account_sid:
          type: string
          format: uuid
        application_sid:
          type: string
          format: uuid
        call_id:
          type: string
        call_sid:
          type: string
          format: uuid
        call_status:
          type: string
          enum:
            - trying
            - ringing
            - alerting
            - in-progress
            - completed
            - busy
            - no-answer
            - failed
            - queued
        caller_name:
          type: string
        direction:
          type: string
          enum:
            - inbound
            - outbound 
        duration:
          type: integer
        from:
          type: string
        originating_sip_trunk_name:
          type: string
        parent_call_sid:
          type: string
          format: uuid
        service_url:
          type: string
        sip_status:
          type: integer
        to:
          type: string
      required:
        - account_sid
        - call_id
        - call_sid
        - call_status
        - direction
        - from
        - service_url
        - sip_status
        - to
    Target:
      properties:
        type:
          type: string
          enum:
            - phone
            - sip
            - user
            - teams
        number:
          type: string
        sipUri:
          type: string
        tenant:
          type: string
          description: Microsoft Teams customer tenant domain name
        trunk:
          type: string
        vmail:
          type: boolean
          description: Dial directly into user's voicemail to leave a message
        overrideTo:
          type: string
        name:
          type: string
        auth:
          type: object
          properties:
            username:
              type: string
            password:
              type: string
      required:
        - type
      example: {"type": "phone", "number": "+16172375080"}
...
    UserProfile:
      type: object
      properties:
        user_sid: 
          type: string
          example: a5bce31e-a028-45cd-94c4-f121b72fec61
        account_sid:
          type: string
          example: a5bce31e-a028-45cd-94c4-f121b72fec61
        is_active:
          type: boolean
          description: indicates whether account is active
          example: true
        name:
          type: string
          description: full user name
          example: Dave Horton
        email:
          type: string
          description: email associated with user
          example: daveh@drachtio.org
        phone:
          type: string
          description: phone associated with user
        provider:
          type: string
          description: authentication provider
          enum:
            - github
            - google
            - local
          example: github
        scope:
          type: string
          description: scope of user permissions
          enum:
            - read-only
            - read-write
          example: read-write
        pristine:
          type: boolean
          description: true if account was newly created
          example: true
        email_validated:
          type: boolean
          description: indicates whether user has validated their email address
          example: true
        phone_validated:
          type: boolean
          description: indicates whether user has validated their mobile phone
          example: true
        tutorial_completion:
          type: number
          description: bitmask indicating which tutorials have been completed
          example: 1
        jwt:
          type: string
          description: json web token to be used as bearer token in API requests
      required:
        - user_sid
        - provider
        - name
        - email
        - scope
        - account_validated
        - pristine
        - jwt
        - is_active
...
security:
  - bearerAuth: []